ai/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
pillow>=9.5.0
numpy<2.0.0,>=1.24.0
opencv-python>=4.8.0
albumentations>=1.3.0
diffusers>=0.19.0
transformers>=4.30.0
accelerate>=0.21.0
scikit-learn>=1.2.2
python-dotenv>=1.0.0
seaborn>=0.12.0
scipy>=1.10.0
